% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Example microbiome dataset bundle}
\format{
A named list with 5 elements: \code{O}, \code{dist_mat}, \code{meta}, \code{taxonomy}, \code{tree}.
}
\usage{
data(exampleData)
}
\description{
A list containing example OTU table, sample metadata, taxonomy, tree, and sequence distance matrix.
}
\keyword{datasets}
