% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData_transfer}
\alias{exampleData_transfer}
\title{Example microbiome dataset bundle for transfer learning}
\format{
A named list file with 2 elements: \code{new_data}, \code{new_meta}.
}
\usage{
data(exampleData_transfer)
}
\description{
A list containing example OTU table, sample metadata.
}
\keyword{datasets}
