% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDICT.R
\name{metadict}
\alias{metadict}
\title{Microbiome data integration method via shared dictionary learning.}
\usage{
metadict(
  O_list,
  dist,
  meta.list = NULL,
  alpha = 1,
  beta = 0.01,
  gamma = 1,
  normalization = "uq",
  max_iter = 10000,
  neighbor = 5,
  sigma = 10
)
}
\arguments{
\item{O_list}{A \code{list} object contains count tables for different datasets. The rows represent taxa, and the columns represent samples. All count tables must have the same taxa.}

\item{dist}{Sequence dissimilarity matrix. Dissimilarity can be measured by phylogenetic distance or taxonomic dissimilarity.}

\item{meta.list}{A \code{list} object contains metadata tables for different datasets.}

\item{alpha}{This parameter is related to the rank of the final corrected count table. A larger alpha leads to a lower rank in the corrected count table.}

\item{beta}{This parameter is related to the smoothness of the estimated measurement efficiency. A larger beta results in more similar measurement efficiency across taxa.}

\item{gamma}{This parameter is related to how well the model fits the data.}

\item{normalization}{Normalization method. Upper quantile or RSim.}

\item{max_iter}{Maximum number of iterations for the optimization process.}

\item{neighbor}{The number of nearest neighbors used to construct the taxa neighbor graph. The default is 5.}

\item{sigma}{Hyperparameter for the Gaussian kernel used to determine graph edge weights.}
}
\value{
a \code{list} with components:
\itemize{
\item{\code{X}, a matrix. Integrated count table. Rows represent taxa. Columns represent samples.}
\item{\code{D}, a matrix. Estimated shared dictionary.}
\item{\code{R}, a matrix. Estimated sample representation.}
\item{\code{w}, a matrix. Estimated measurement efficiency. Rows represent dataset, columns represent taxa.}
\item{\code{meta}, a data frame. Meta table without batch ID. Sample features must be present in all datasets. Default is NULL.}
}
}
\description{
A method for the integration of microbiome data. This method is designed to remove batch effects and preserve biological variation while integrating heterogeneous datasets. MetaDICT can better avoid overcorrection when unobserved confounding variables are present.
}
\details{
MetaDICT is a two-step approach. It initially estimates the batch effects by weighting methods in causal inference literature then refines the estimation via a shared dictionary learning.
}
\examples{

 data(exampleData)
 alpha = 0.01
 beta = 0.01
 gamma = 1
 metadict_res = metadict(O_list,dist,meta.list = meta.list)

}
