% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_function.R
\name{pcoa.plot.continuous}
\alias{pcoa.plot.continuous}
\title{PCoA plots for continuous variables.}
\usage{
pcoa.plot.continuous(
  X,
  covariate,
  title,
  R2 = TRUE,
  dissimilarity = "Bray-Curtis",
  point_size = 1
)
}
\arguments{
\item{X}{Abundance matrix. The rows represent taxa, the columns represent samples.}

\item{covariate}{A discrete sample covariate.}

\item{title}{Graph title.}

\item{R2}{A logical variable. Whether to show R2 statistic in the subtitle. Default is TRUE.}

\item{dissimilarity}{The dissimilarity type, “Bray-Curtis” for Bray-Curtis dissimilarity,
“Euclidean” for generalized UniFrac dissimilarity.}

\item{point_size}{Point size. Default is 1.}
}
\value{
a PCoA plot.
}
\description{
PCoA plots for continuous variables.
}
