% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_function.R
\name{community_detection}
\alias{community_detection}
\title{Taxa/Sample Community detection.}
\usage{
community_detection(
  X,
  max_k = 10,
  method = "Louvain",
  resolution = 1,
  min_k = 2
)
}
\arguments{
\item{X}{Input data. The rows represent clustering objects, the columns represent features.}

\item{max_k}{The largest number of connected neighbors.}

\item{method}{Community detection methods such as Louvain and Walktrap.}

\item{resolution}{The resolution parameter of Louvain algorithm.}

\item{min_k}{The smallest number of connected neighbors.}
}
\value{
a \code{list} with components:
\itemize{
\item{\code{cluster}, the estimated cluster labels.}
\item{\code{graph}, the \code{k}-nearest neighbor graph.}
}
}
\description{
A \code{k}-nearest neighbor graph is constructed based on Euclidean distance. Then community detection method is applied to identify communities.
Various values of \code{k} within a specific range are tried and the one that yields the highest average Silhouette score is selected.
}
