% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDICT.R
\name{MetaDICT}
\alias{MetaDICT}
\title{Microbiome data integration method via shared dictionary learning.}
\usage{
MetaDICT(
  count,
  meta,
  covariates = "all",
  tree = NULL,
  taxonomy = NULL,
  distance_matrix = NULL,
  tax_level = NULL,
  customize_parameter = FALSE,
  alpha = 0.1,
  beta = 0.01,
  normalization = "uq",
  max_iter = 10000,
  verbose = TRUE,
  optim_trace = FALSE
)
}
\arguments{
\item{count}{The integrated count table of taxa by samples. The
\code{count} parameter should be provided as either a \code{matrix} or a
\code{data.frame}.}

\item{meta}{The integrated meta table \code{meta} contains sample information
and batch id.}

\item{covariates}{The covariates used in data integration. Default is all.}

\item{tree}{The phylogenetic tree (optional if distance matrix or taxonomy is provided).}

\item{taxonomy}{The taxonomy table (optional if distance matrix or phylogenetic tree is provided).}

\item{distance_matrix}{The \code{matrix} that measures the dismilarity of taxa. Default is NULL and MetaDICT generates
distance matrix based on phylogenetic tree and taxonomic information.}

\item{tax_level}{The taxonomic level of count table.}

\item{customize_parameter}{A logical variable. TRUE if the parameters of alpha and beta
are customized. FALSE if the parameters are determined by MetaDICT based on the number
of covariates provided.}

\item{alpha}{This parameter is related to the rank of the final corrected count table.
A larger alpha leads to a lower rank of shared dictionary.}

\item{beta}{This parameter is related to the smoothness of the estimated measurement efficiency.
A larger beta results in more similar measurement efficiency across taxa.}

\item{normalization}{Normalization method. Upper quantile or RSim. Set to be NULL if
normalization is not needed.}

\item{max_iter}{Maximum number of iterations for the optimization process. Default is 10000.}

\item{verbose}{A logical variable. Wheter to generate verbose output. Default is TRUE.}

\item{optim_trace}{A logical variable. Whether to print the optimization step. Default is FALSE.}
}
\value{
a \code{list} with components:
\itemize{
\item{\code{count}, a \code{data.frame}. Corrected count table. Rows represent taxa. Columns represent samples.}
\item{\code{D}, a \code{matrix}. Estimated shared dictionary.}
\item{\code{R}, a \code{matrix}. List of estimated sample representation.}
\item{\code{w}, a \code{matrix}. Estimated measurement efficiency. Rows represent dataset, columns represent taxa.}
\item{\code{meta}, a \code{data.frame}. Meta table without batch ID. Sample features must be present in all datasets. Default is NULL.}
}
}
\description{
A method for microbiome data integration. This method is designed to
remove batch effects and preserve biological variation while integrating heterogeneous datasets.
MetaDICT can better avoid overcorrection when unobserved confounding variables are present.
}
\details{
MetaDICT is a two-step approach. It initially estimates the batch effects by covariate balancing,
then refines the estimation via shared dictionary learning.
}
\examples{
 data(exampleData)
 metadict_res = metadict(O, meta, distance_matrix = dist_mat, alpha = 0.01, beta = 0.01)

}
