% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_function.R
\name{pcoa.plot.continuous}
\alias{pcoa.plot.continuous}
\title{PCoA plots for continuous variable with bray-curtis distance.}
\usage{
pcoa.plot.continuous(A, covariate, main)
}
\arguments{
\item{A}{Abundance matrix. The rows represent taxa, the columns represent samples.}

\item{covariate}{A continuous sample covariate.}

\item{main}{Graph title.}
}
\value{
a PCoA plot.
}
\description{
PCoA plots for continuous variable with bray-curtis distance.
}
